+++ vcxproj2cmake.Tests/ConverterTests/TranslateMSBuildMacrosTests.cs
@@ -23,6 +23,9 @@ public partial class ConverterTests
                     "$(ProJECtName).cpp",
                     "$(SolUTIonDir)SomeFile.cpp",
                     "$(SolUTIonName).cpp")
+                .WithClCompileSetting(
+                    "PreprocessorDefinitions",
+                    "OUT_DIR=$(OutDir);TARGET_DIR=$(TargetDir);TARGET_EXT=$(TargetExt);TARGET_FILE_NAME=$(TargetFileName);TARGET_NAME=$(TargetName);TARGET_PATH=$(TargetPath)")
                 .Build()));
 
             var converter = new Converter(fileSystem, NullLogger.Instance);
@@ -36,12 +39,23 @@ public partial class ConverterTests
             Assert.Contains("""
                 target_sources(Project
                     PRIVATE
-                        "${CMAKE_BUILD_TYPE}.cpp"
-                        "${CMAKE_BUILD_TYPE}.cpp"
                         "${CMAKE_CURRENT_SOURCE_DIR}/SomeFile.cpp"
                         "${CMAKE_PROJECT_NAME}.cpp"
                         "${CMAKE_SOURCE_DIR}/SomeFile.cpp"
                         "${PROJECT_NAME}.cpp"
+                        "$<CONFIG>.cpp"
+                        "$<CONFIG>.cpp"
+                )
+                """, cmake);
+            Assert.Contains("""
+                target_compile_definitions(Project
+                    PRIVATE
+                        "OUT_DIR=$<TARGET_FILE_DIR:Project>/"
+                        "TARGET_DIR=$<TARGET_FILE_DIR:Project>/"
+                        "TARGET_EXT=$<TARGET_FILE_SUFFIX:Project>"
+                        "TARGET_FILE_NAME=$<TARGET_FILE_NAME:Project>"
+                        "TARGET_NAME=$<TARGET_FILE_BASE_NAME:Project>"
+                        "TARGET_PATH=$<TARGET_FILE:Project>"
                 )
                 """, cmake);
         }