+++ vcxproj2cmake/CMakeGenerator.cs
@@ -182,8 +182,9 @@ class CMakeGenerator
         var path = normalizedPath.Value;
         var isAbsolutePath = Path.IsPathRooted(path);
 
-        // if a path starts with a CMake variable, we just assume that the variable resolves to an absolute path
+        // if a path starts with a CMake variable or generator expression, we just assume that it resolves to an absolute path
         isAbsolutePath |= path.StartsWith("${");
+        isAbsolutePath |= path.StartsWith("$<");
 
         if (!isAbsolutePath)
             if (path == ".")